chore: add config to set permission background & move watermark ui to debugger package
