+++ packages-experimental/debugger/src/controllers/menu.schema.ts
@@ -24,6 +24,7 @@ import { DialogOperation } from '../commands/operations/dialog.operation';
 import { LocaleOperation } from '../commands/operations/locale.operation';
 import { MessageOperation } from '../commands/operations/message.operation';
 import { NotificationOperation } from '../commands/operations/notification.operation';
+import { OpenWatermarkPanelOperation } from '../commands/operations/open-watermark-panel.operation';
 import { SaveSnapshotOptions } from '../commands/operations/save-snapshot.operations';
 import { SetEditable } from '../commands/operations/set.editable.operation';
 import { SidebarOperation } from '../commands/operations/sidebar.operation';
@@ -49,6 +50,7 @@ import {
     ThemeMenuItemFactory,
     UNIT_ITEM_MENU_ID,
     UnitMenuItemFactory,
+    WatermarkMenuItemFactory,
 } from '../controllers/menu';
 
 export const menuSchema: MenuSchemaType = {
@@ -124,6 +126,10 @@ export const menuSchema: MenuSchemaType = {
         [ChangeUserCommand.id]: {
             order: 11,
             menuItemFactory: ChangeUserMenuItemFactory,
+        },
+        [OpenWatermarkPanelOperation.id]: {
+            order: 12,
+            menuItemFactory: WatermarkMenuItemFactory,
         },
     },
 };
