+++ packages-experimental/debugger/src/debugger-plugin.ts
@@ -21,6 +21,7 @@ import { defaultPluginConfig, PLUGIN_CONFIG_KEY } from './controllers/config.sch
 import { DebuggerController } from './controllers/debugger.controller';
 import { E2EController } from './controllers/e2e/e2e.controller';
 import { PerformanceMonitorController } from './controllers/performance-monitor.controller';
+import { UniverWatermarkMenuController } from './controllers/watermark.menu.controller';
 
 export class UniverDebuggerPlugin extends Plugin {
     static override pluginName = 'UNIVER_DEBUGGER_PLUGIN';
@@ -47,6 +48,7 @@ export class UniverDebuggerPlugin extends Plugin {
             [PerformanceMonitorController],
             [DebuggerController],
             [E2EController],
+            [UniverWatermarkMenuController],
         ] as Dependency[]).forEach((d) => this._injector.add(d));
 
         this._injector.get(E2EController);
@@ -58,6 +60,7 @@ export class UniverDebuggerPlugin extends Plugin {
 
     override onRendered(): void {
         this._injector.get(PerformanceMonitorController);
+        this._injector.get(UniverWatermarkMenuController);
     }
 
     getDebuggerController() {
